backend/api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
aiokafka
pydantic
python-dotenv
kafka-python